requirements.txt
# Requirements file installation
matplotlib
numpy
django
djangorestframework
django-cors-headers
coreapi
pytest